# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  provider: string;
  detail: string;
}

export interface ProjectInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export type ProjectReportStatus = typeof ProjectReportStatus[keyof typeof ProjectReportStatus];


export const ProjectReportStatus = {
  not_started: 'not_started',
  ready: 'ready',
} as const;

export interface Project {
  id: string;
  title: string;
  createdAt: string;
  assetCount: number;
  detectionCount: number;
  reportStatus: ProjectReportStatus;
}

export type AssetInputType = typeof AssetInputType[keyof typeof AssetInputType];


export const AssetInputType = {
  script: 'script',
  image: 'image',
  video: 'video',
} as const;

export interface AssetInput {
  filename: string;
  type: AssetInputType;
  mimeType: string;
  /** @minLength 1 */
  contentBase64: string;
  width?: number;
  height?: number;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  script: 'script',
  image: 'image',
  video: 'video',
} as const;

export interface Asset {
  id: string;
  projectId: string;
  filename: string;
  type: AssetType;
  mimeType: string;
  uploadedAt: string;
  sizeBytes: number;
  width: number;
  height: number;
}

export interface BoundingBox {
  left: number;
  top: number;
  right: number;
  bottom: number;
}

export type AssetPreviewType = typeof AssetPreviewType[keyof typeof AssetPreviewType];


export const AssetPreviewType = {
  image: 'image',
  video: 'video',
} as const;

export interface AssetPreview {
  assetId: string;
  filename: string;
  type: AssetPreviewType;
  mimeType: string;
  dataUrl: string;
  width: number;
  height: number;
}

export type DetectionCategory = typeof DetectionCategory[keyof typeof DetectionCategory];


export const DetectionCategory = {
  brand: 'brand',
  logo: 'logo',
  celebrity_name: 'celebrity_name',
  song: 'song',
  existing_ip: 'existing_ip',
} as const;

export type DetectionRiskLevel = typeof DetectionRiskLevel[keyof typeof DetectionRiskLevel];


export const DetectionRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

/**
 * @nullable
 */
export type DetectionProminence = typeof DetectionProminence[keyof typeof DetectionProminence] | null;


export const DetectionProminence = {
  background: 'background',
  featured: 'featured',
} as const;

export interface Detection {
  id: string;
  assetId: string;
  category: DetectionCategory;
  name: string;
  sourceRef: string;
  contextSnippet: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  riskLevel: DetectionRiskLevel;
  rationale: string;
  boundingBox: BoundingBox | null;
  /** @nullable */
  frameReference: string | null;
  /** @nullable */
  prominence: DetectionProminence;
}

export interface ReportCounts {
  low: number;
  medium: number;
  high: number;
}

export interface Report {
  projectId: string;
  generatedAt: string;
  summary: string;
  counts: ReportCounts;
  detections: Detection[];
  analyzedAssets: number;
  previews: AssetPreview[];
}