# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clearance
    description: Rights clearance scanning operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [clearance]
      summary: List clearance projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [clearance]
      summary: Create a clearance project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}/assets:
    post:
      operationId: uploadAsset
      tags: [clearance]
      summary: Add a script, image, or video asset
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Uploaded asset metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Project not found
  /projects/{projectId}/analyze:
    post:
      operationId: analyzeProject
      tags: [clearance]
      summary: Analyze all project assets with Gemini
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Compiled clearance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Project has no assets
        "404":
          description: Project not found
  /projects/{projectId}/report:
    get:
      operationId: getProjectReport
      tags: [clearance]
      summary: Get the latest clearance report
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Latest clearance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        provider:
          type: string
        detail:
          type: string
      required: [status, provider, detail]
    ProjectInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
      required:
        - title
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
        assetCount:
          type: number
        detectionCount:
          type: number
        reportStatus:
          type: string
          enum: [not_started, ready]
      required:
        - id
        - title
        - createdAt
        - assetCount
        - detectionCount
        - reportStatus
    AssetInput:
      type: object
      properties:
        filename:
          type: string
        type:
          type: string
          enum: [script, image, video]
        mimeType:
          type: string
        contentBase64:
          type: string
          minLength: 1
        width:
          type: number
        height:
          type: number
      required:
        - filename
        - type
        - mimeType
        - contentBase64
    Asset:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [script, image, video]
        mimeType:
          type: string
        uploadedAt:
          type: string
        sizeBytes:
          type: number
        width:
          type: number
        height:
          type: number
      required:
        - id
        - projectId
        - filename
        - type
        - mimeType
        - uploadedAt
        - sizeBytes
        - width
        - height
    BoundingBox:
      type: object
      properties:
        left:
          type: number
        top:
          type: number
        right:
          type: number
        bottom:
          type: number
      required: [left, top, right, bottom]
    AssetPreview:
      type: object
      properties:
        assetId:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [image, video]
        mimeType:
          type: string
        dataUrl:
          type: string
        width:
          type: number
        height:
          type: number
      required:
        - assetId
        - filename
        - type
        - mimeType
        - dataUrl
        - width
        - height
    Detection:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
        category:
          type: string
          enum: [brand, logo, celebrity_name, song, existing_ip]
        name:
          type: string
        sourceRef:
          type: string
        contextSnippet:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        riskLevel:
          type: string
          enum: [low, medium, high]
        rationale:
          type: string
        boundingBox:
          anyOf:
            - $ref: "#/components/schemas/BoundingBox"
            - type: "null"
        frameReference:
          type:
            - string
            - "null"
        prominence:
          type:
            - string
            - "null"
          enum: [background, featured, null]
      required:
        - id
        - assetId
        - category
        - name
        - sourceRef
        - contextSnippet
        - confidence
        - riskLevel
        - rationale
        - boundingBox
        - frameReference
        - prominence
    ReportCounts:
      type: object
      properties:
        low:
          type: number
        medium:
          type: number
        high:
          type: number
      required: [low, medium, high]
    Report:
      type: object
      properties:
        projectId:
          type: string
        generatedAt:
          type: string
        summary:
          type: string
        counts:
          $ref: "#/components/schemas/ReportCounts"
        detections:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
        analyzedAssets:
          type: number
        previews:
          type: array
          items:
            $ref: "#/components/schemas/AssetPreview"
      required:
        - projectId
        - generatedAt
        - summary
        - counts
        - detections
        - analyzedAssets
        - previews