# lib/api-zod/src/generated/types/assetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetInputType } from './assetInputType';

export interface AssetInput {
  filename: string;
  type: AssetInputType;
  mimeType: string;
  /** @minLength 1 */
  contentBase64: string;
  width?: number;
  height?: number;
}