# lib/api-zod/src/generated/types/assetPreview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetPreviewType } from './assetPreviewType';

export interface AssetPreview {
  assetId: string;
  filename: string;
  type: AssetPreviewType;
  mimeType: string;
  dataUrl: string;
  width: number;
  height: number;
}