# lib/api-zod/src/generated/types/boundingBox.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoundingBox {
  left: number;
  top: number;
  right: number;
  bottom: number;
}