# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoundingBox } from './boundingBox';
import type { DetectionCategory } from './detectionCategory';
import type { DetectionProminence } from './detectionProminence';
import type { DetectionRiskLevel } from './detectionRiskLevel';

export interface Detection {
  id: string;
  assetId: string;
  category: DetectionCategory;
  name: string;
  sourceRef: string;
  contextSnippet: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  riskLevel: DetectionRiskLevel;
  rationale: string;
  boundingBox: BoundingBox | null;
  /** @nullable */
  frameReference: string | null;
  /** @nullable */
  prominence: DetectionProminence;
}