# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetPreview } from './assetPreview';
import type { Detection } from './detection';
import type { ReportCounts } from './reportCounts';

export interface Report {
  projectId: string;
  generatedAt: string;
  summary: string;
  counts: ReportCounts;
  detections: Detection[];
  analyzedAssets: number;
  previews: AssetPreview[];
}